requirements.txt
amqp==2.1.4
anyjson==0.3.3
billiard==3.5.0.2
celery==4.0.2
Django==1.11
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-celery-statuspage==0.0.3
django-contrib-comments==1.8.0
django-cuser==2017.3.16
django-material==0.13.1
django-redis==4.7.0
kombu==4.0.2
pytz==2017.2
redis==2.10.5
requests==2.13.0
titlecase==0.9.0
vine==1.1.3
